pyproject.toml
[project]
name = "techchallenge-artefact"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain==0.2.16",
    "langchain-community==0.2.16",
    "langchain-core==0.2.38",
    "langchain-text-splitters==0.2.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.14.11",
]
[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]
